Add command to deploy the package
Adding a new command to deploy a postgreSQL project. User can right click on the pgproj file in editor or explorer and also run the command via command palette.
The command will first build the project, if successful it will try and find the file to deploy which will be projectName.sql (We do this as we do not know the path of the script file) then it opens the connectionDialog with only PostgreSQL and try to deploy the script if connection was successful.